CACTUS-102 - initial sunset deprecated Sortable adapter in favor of Freestar sans Native/Video
