--- modules/sortableAnalyticsAdapter.md
@@ -1,9 +1,0 @@
-# Overview
-
-Module Name: Sortable Analytics Adapter
-Module Type: Analytics Adapter
-Maintainer: prebid@sortable.com
-
-# Description
-
-Analytics adapter for Sortable. Contact prebid@sortable.com for information.
